+++ libs/open-api/remote-method/src/lib/open-api.remote-method.ts
@@ -15,7 +15,9 @@ import {
   OpenApiConfigService,
   SchemaValidationMixin,
   RXAP_OPEN_API_STRICT_VALIDATOR,
-  DEFAULT_OPEN_API_REMOTE_METHOD_META_DATA
+  DEFAULT_OPEN_API_REMOTE_METHOD_META_DATA,
+  DISABLE_SCHEMA_VALIDATION,
+  DISABLE_VALIDATION
 } from '@rxap/open-api';
 import { BaseHttpRemoteMethod } from '@rxap/remote-method/http';
 import {
@@ -70,6 +72,8 @@ export class OpenApiRemoteMethod<Response = any, Parameters extends Record<strin
   }
 
   private readonly strict: boolean = false;
+  protected readonly disableSchemaValidation: boolean = false;
+  protected readonly disableValidation: boolean = false;
 
   constructor(
     @Inject(HttpClient) http: HttpClient,
@@ -80,7 +84,13 @@ export class OpenApiRemoteMethod<Response = any, Parameters extends Record<strin
       metadata: OpenApiRemoteMethodMetadata | null = null,
     @Optional()
     @Inject(RXAP_OPEN_API_STRICT_VALIDATOR)
-      strict: boolean | null                       = null
+      strict: boolean | null                                   = null,
+    @Optional()
+    @Inject(DISABLE_SCHEMA_VALIDATION)
+    disableSchemaValidation: boolean | null = null,
+    @Optional()
+    @Inject(DISABLE_VALIDATION)
+      disableValidation: boolean | null = null
   ) {
     super(http, injector, metadata);
     let operation: OperationObjectWithMetadata;
@@ -97,12 +107,16 @@ export class OpenApiRemoteMethod<Response = any, Parameters extends Record<strin
       ignoreUndefined: this.metadata.ignoreUndefined ?? true,
     });
     this.strict = strict || this.metadata.strict || false;
+    this.disableSchemaValidation = disableSchemaValidation || this.metadata.disableSchemaValidation || false;
+    this.disableValidation = disableValidation || this.metadata.disableValidation || false;
   }
 
   protected async _call(args: OpenApiRemoteMethodParameter<Parameters, RequestBody> = {}): Promise<Response> {
 
-    this.validateParameters(this.operation, args.parameters, this.strict);
-    this.validateRequestBody(this.operation, args.requestBody, this.strict);
+    if (!this.disableValidation) {
+      this.validateParameters(this.operation, args.parameters, this.strict);
+      this.validateRequestBody(this.operation, args.requestBody, this.strict);
+    }
 
     const response = await this._callWithResponse(args);
 
@@ -132,8 +146,10 @@ export class OpenApiRemoteMethod<Response = any, Parameters extends Record<strin
   // TODO : update to the new call method concept (the remove of the _call method concept)
   public _callWithResponse(args: OpenApiRemoteMethodParameter<Parameters, RequestBody> = {}): Promise<HttpResponse<Response>> {
 
-    this.validateParameters(this.operation, args.parameters, this.strict);
-    this.validateRequestBody(this.operation, args.requestBody, this.strict);
+    if (!this.disableValidation) {
+      this.validateParameters(this.operation, args.parameters, this.strict);
+      this.validateRequestBody(this.operation, args.requestBody, this.strict);
+    }
 
     return this.http.request<Response>(this.updateRequest(this.buildHttpOptions(this.operation, args.parameters, args.requestBody, this.metadata.ignoreUndefined))).pipe(
       retry(this.metadata.retry ?? 0),
@@ -144,7 +160,11 @@ export class OpenApiRemoteMethod<Response = any, Parameters extends Record<strin
         return throwError(response);
       }),
       filter((event: any) => event.type === HttpEventType.Response),
-      tap((response: HttpResponse<Response>) => this.validateResponse(this.operation, response, this.strict)),
+      tap((response: HttpResponse<Response>) => {
+        if (this.disableValidation) {
+          this.validateResponse(this.operation, response, this.strict);
+        }
+      }),
       tap((response: HttpResponse<Response>) => {
         this.interceptors?.forEach(interceptor => interceptor
           .next({ response, parameters: args.parameters, requestBody: args.requestBody })