+++ libs/oauth/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [12.1.0](https://gitlab.com/rxap/packages/compare/@rxap/oauth@12.0.11...@rxap/oauth@12.1.0) (2022-03-25)
+
+
+### Features
+
+* add auth event system ([b421849](https://gitlab.com/rxap/packages/commit/b421849721c7b76bb7963d0fd440943e25779f5c))
+
+
+
+
+
+## [12.0.11](https://gitlab.com/rxap/packages/compare/@rxap/oauth@12.0.10...@rxap/oauth@12.0.11) (2022-03-25)
+
+
+### Bug Fixes
+
+* support clientId and clientSecret auth flow ([e17381c](https://gitlab.com/rxap/packages/commit/e17381cd48e7189c68a7c03d9cb025cd53c96dc7))
+
+
+
+
+
 # [13.0.0-next.2](https://gitlab.com/rxap/packages/compare/@rxap/oauth@13.0.0-next.1...@rxap/oauth@13.0.0-next.2) (2022-03-29)
 
 **Note:** Version bump only for package @rxap/oauth