+++ libs/open-api/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [12.4.0](https://gitlab.com/rxap/packages/compare/@rxap/open-api@12.3.0...@rxap/open-api@12.4.0) (2022-04-06)
+
+
+### Features
+
+* add injection token to disable validation ([adf80ad](https://gitlab.com/rxap/packages/commit/adf80adbde0b039d4ef33e60205b625ca180706f))
+
+
+
+
+
+# [12.3.0](https://gitlab.com/rxap/packages/compare/@rxap/open-api@12.2.7...@rxap/open-api@12.3.0) (2022-04-06)
+
+
+### Features
+
+* add injection token to disable schema validation ([668ea74](https://gitlab.com/rxap/packages/commit/668ea742ba88e67f7dd967797e4ef179d0b389f7))
+
+
+
+
+
 # [13.0.0-next.2](https://gitlab.com/rxap/packages/compare/@rxap/open-api@13.0.0-next.1...@rxap/open-api@13.0.0-next.2) (2022-03-29)
 
 **Note:** Version bump only for package @rxap/open-api